Outh only logic

+++ app/Http/Controllers/IdeaController.php
@@ -27,6 +27,10 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user->id) {
+            abort(404);
+        }
+
         $editing = true;
 
         return view('ideas.show', compact('idea', 'editing'));
@@ -34,6 +38,10 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
+        if (auth()->id() !== $idea->user->id) {
+            abort(404);
+        }
+
         $validated = request()->validate([
             'content' => 'required|min:5|max:240'
         ]);
@@ -45,6 +53,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user->id) {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', 'Idea deleted successfully.');
